feat: comprehensive combat, boss, and upgrade system overhaul
- **Major Upgrade System Overhaul**: Replaced the legacy shop with a 3-branch `Decompiler v4.0` (Architect, Enforcer, Ghost). Introduced new mechanics including passive HP regeneration, critical hit exploits, kinetic overloads on kill, and multi-shot projectiles.
- **Advanced Boss Behaviors**: Implemented a specialized behavior system for five distinct bosses (`Guardian`, `Stalker`, `Glitch`, `Omega`, and the level-100 `Ultima`). Bosses now feature unique attack patterns like seismic jumps, projectile throws, teleportation, and phase-based multi-attacks.
- **Hybrid Objective System**: Added support for overlapping gameplay modes, such as the new Defense+Escape hybrid where players must locate and protect a core before extracting.
- **Improved Player AI**: Enhanced Autopilot with "Weapon-Aware Kiting" (dynamic distancing based on equipped weapon), auto-deployment of Sonar in darkness, and smart defensive trap usage.
- **Engine & Performance**:
    - Migrated score and data tracking to `Long` types to prevent overflow during high-level play.
    - Implemented "Saturated Growth Curves" for smoother scaling of HP, rewards, and difficulty up to level 200.
    - Improved AABB collision sliding and added enemy projectile detection.
- **Technical Refactoring**:
    - Updated `compileSdk` to 37 and incremented version to `0.7.2`.
    - Integrated comprehensive unit and instrumented tests for level simulation and maze connectivity.
    - Optimized `SaveManager` with safe migration logic for legacy integer data.
- **UI Enhancements**: Refined `UINanoOS` and `UIDecompiler` layouts for better readability in both portrait and landscape modes.

+++ app/src/main/java/com/appsbyalok/echohunter/data/StoryProtocol.kt
@@ -65,8 +65,8 @@ object StoryProtocol {
         R.string.story_bad_1, R.string.story_bad_2, R.string.story_bad_3
     )
 
-    fun triggerRandomGlitch(score: Int, gameMode: Int, difficulty: Int) {
-        if (score > 150 && Random.nextDouble() < 0.05) {
+    fun triggerRandomGlitch(score: Long, gameMode: Int, difficulty: Int) {
+        if (score > 150L && Random.nextDouble() < 0.05) {
             isGlitchActive = true
             showIngameMessage("ADMIN: \"I FOUND YOUR IP.\"", 2f)
         }
